refactor: 新增数据结构目录
